pyproject.toml
[tool.poetry]
name = "4tochki-parser"
version = "0.1.0"
description = ""
authors = ["a_cheremisin <a_cheremisin@lesta.group>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
structlog = "^25.5.0"
python-decouple = "^3.8"
redis = "^7.2.0"
requests = "^2.32.5"
pydantic = "^2.12.5"
sqlalchemy = "^2.0.47"
pymysql = "^1.1.2"
beautifulsoup4 = "^4.14.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
